refactor open interactive
